requirements.txt
streamlit==1.20.0
opencv-python-headless==4.6.0.66
numpy==1.24.0
Pillow==9.4.0
pickle-mixin==1.0.2